+++ src/common/index.ts
@@ -1,5 +1,4 @@
 export * from './buffered-writable';
 export * from './constructor';
-export * from './discriminant';
 export * from './writable';
 export * from './incomplete-read-result';